Update Travis, language->implementation, tidy up testing framework
