+++ .travis.yml
@@ -1,8 +1,5 @@
 language: node_js
 node_js:
   - "4.1"
-  - "4.0"
   - "0.12"
-  - "0.11"
-  - "0.10"
   - "iojs"
